+++ Client/Frontend/Home/LogoHeader/FxHomeLogoHeaderCell.swift
@@ -15,7 +15,7 @@ class FxHomeLogoHeaderCell: UICollectionViewCell, ReusableCell {
     typealias a11y = AccessibilityIdentifiers.FirefoxHomepage.OtherButtons
 
     // MARK: - UI Elements
-    lazy var logoButton: UIButton = .build { button in
+    lazy var logoButton: ActionButton = .build { button in
         button.setTitle("", for: .normal)
         button.backgroundColor = .clear
         button.accessibilityIdentifier = a11y.logoButton
@@ -55,6 +55,10 @@ class FxHomeLogoHeaderCell: UICollectionViewCell, ReusableCell {
             logoButton.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
             logoButton.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -10)
         ])
+    }
+
+    func configure(onTapAction: ((UIButton) -> Void)?) {
+        logoButton.touchUpAction = onTapAction
     }
 
     // MARK: - Animation
